backend/requirements.txt
# --- Core ---
fastapi>=0.115
uvicorn[standard]>=0.34
python-dotenv>=1.0
pydantic>=2.10
python-multipart>=0.0.20

# --- AI / LangChain ---
langchain>=0.3
langchain-openai>=0.3
langgraph>=0.2.60
langchain-core>=0.3
openai>=1.58

# --- Utilities ---
httpx>=0.28
